Use long for sendfile size.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jni@300516 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/tomcat/jni/Socket.java
@@ -441,8 +441,8 @@ public class Socket {
      *         file, and trailers
      *
      */
-    public static native int sendfile(long sock, long file, byte [][] headers,
-                                      byte[][] trailers, long offset,
-                                      int len, int flags);
+    public static native long sendfile(long sock, long file, byte [][] headers,
+                                       byte[][] trailers, long offset,
+                                       long len, int flags);
 
 }
